+++ src/main/java/parser/csv/CsvOutputStream.java
@@ -1,4 +1,4 @@
-package parser;
+package parser.csv;
 
 import java.io.File;
 import java.io.IOException;
